+++ src/main/java/org/folio/s3/client/FolioS3Client.java
@@ -4,6 +4,7 @@ import io.minio.http.Method;
 
 import java.io.InputStream;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 
 public interface FolioS3Client {
   /**
@@ -44,6 +45,16 @@ public interface FolioS3Client {
   String write(String path, InputStream is, long size);
 
   /**
+   * Composes multiple objects into a single object. Note that each object
+   * except the last must be at least 5MB (minimum multipart part size)
+   *
+   * @param destination the destination object key
+   * @param sourceKeys the list of source object keys to compose
+   * @return the newly created composed object key
+   */
+  String compose(String destination, List<String> sourceKeys);
+
+  /**
    * Removes a file on S3 storage
    *
    * @param path the path to the file to delete
@@ -75,6 +86,14 @@ public interface FolioS3Client {
    * @return list of object paths
    */
   List<String> list(String path);
+
+  /**
+   * Get list of object paths, recursively
+   *
+   * @param path - the path to the file on S3-compatible storage
+   * @return list of object paths
+   */
+  List<String> listRecursive(String path);
 
   /**
    * Get iterable list of object paths
@@ -117,6 +136,17 @@ public interface FolioS3Client {
    * @return presigned url of object
    */
   String getPresignedUrl(String path, Method method);
+
+  /**
+   * Returns presigned url for object on S3-compatible storage with custom expiry
+   *
+   * @param path       - the path to the file on S3-compatible storage
+   * @param method     - http method
+   * @param expiryTime - expiry time
+   * @param expiryUnit - expiry time unit
+   * @return presigned url of object
+   */
+  String getPresignedUrl(String path, Method method, int expiryTime, TimeUnit expiryUnit);
 
   /**
    * Creates bucket. Bucket name should be declared in {@link S3ClientProperties}