+++ NEWS.md
@@ -1,4 +1,10 @@
-## v2.3.2 - In Progress
+## v2.4.0 - In Progress
+
+### Stories
+* [FOLS3CL-40](https://folio-org.atlassian.net/browse/FOLS3CL-40) Add compose, listRecursive, custom presigned expiry
+
+## v2.3.0 - Released
+This release contains upgrade to Java 21 and dependencies upgrading
 
 ## v2.2.2 - Released
 This release contains security fixes for minio library